# Conflicts:
#	core-backend/sandbox/src/funcs.rs
#	core-backend/wasmi/src/funcs/mod.rs
#	gcore/src/exec.rs
#	gstd/src/exec.rs
#	program/src/api/generated/metadata.rs
#	runtime/gear/src/lib.rs
#	runtime/vara/src/lib.rs
#	scripts/test-utils/gear_runtime.compact.compressed.wasm
#	utils/gear-runtime-test-cli/src/command.rs

+++ gear-test/src/proc.rs
@@ -195,7 +195,7 @@ where
                     init_source,
                     init_message.try_into().unwrap(),
                     program.init_gas_limit.unwrap_or(GAS_LIMIT),
-                    program.init_value.unwrap_or(0) as u128,
+                    program.init_value.unwrap_or(0),
                     None,
                 ),
             },
